+++ routes/index.js
@@ -9,7 +9,6 @@ const userRouter = require('./users.js');
 const articlesRouter = require('./articles.js');
 
 app.use(requestLogger);
-
 router.use('/users', userRouter);
 router.use('/articles', articlesRouter);
