--- utils/Errors/configs/config.js
@@ -1,5 +1,0 @@
-const JWT_SECRET = 'EAT_THIS_FISH';
-
-module.exports = {
-  JWT_SECRET,
-};
